after_term support for Javascript highlighter
This allows us to parse / / regex literals correctly in (nearly?) all
situations by recognizing if a literal is currently allowed or not (after terms!).

Now calc(a / b) works perfectly since 'a' is a term.
But match(/blub/) works too since after a ( everything is allowed.

Tested on JQuery 1.6.2 normal source ;-)