Ruby: disable mechanism to find |proc_arg| because of buggy site effects
It is nearly impossible to detect |proc_arg| and to distinguish it from e.g. a | b | c
as an expression - currently the proc_arg hiliter assumes that there are no spaces around
the proc_arg which is too narrow. Yesterday I encountered a buggy site effect that |i1| the
1 is hilited like a number... so I should disable this feature which is currently not enabled
by default completely.